requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
llama-cpp-python>=0.3.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
huggingface-hub>=0.25.0
python-dotenv>=1.0.0
sse-starlette>=2.0.0